1. common.go: now ConvertToInt() and ConvertToUint() function returns int and uint type 2. middleware.go: GetUint(), GetUintByName, GetInt(), GetIntByName() returns uint and int type 3. result of mysql and clickhouse implemented the changes of result interface in middleware.go

+++ middleware/middleware.go
@@ -196,13 +196,13 @@ type Result interface {
 	// IsNullByName checks if value of given row number and column name is nil
 	IsNullByName(row int, name string) (bool, error)
 	// GetUint returns uint64 type value of given row and column number
-	GetUint(row, column int) (uint64, error)
+	GetUint(row, column int) (uint, error)
 	// GetUintByName returns uint64 type value of given row number and column name
-	GetUintByName(row int, name string) (uint64, error)
+	GetUintByName(row int, name string) (uint, error)
 	// GetInt returns int64 type value of given row and column number
-	GetInt(row, column int) (int64, error)
+	GetInt(row, column int) (int, error)
 	// GetIntByName returns int64 type value of given row number and column name
-	GetIntByName(row int, name string) (int64, error)
+	GetIntByName(row int, name string) (int, error)
 	// GetFloat returns float64 type value of given row and column number
 	GetFloat(row, column int) (float64, error)
 	// GetFloatByName returns float64 type value of given row number and column name
